+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Give access to `self` in `TreeForm` click events.
+* Allow `TreeForm` columns to be editable.
 
 ### Removed